# docs/api/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RKM Cinema API",
    "version": "2.0",
    "description": " (FROZEN v1 contract \u2014 see docs/adr/ADR-0001; additive-only)"
  },
  "paths": {
...
    "/api/library/folders/{folder_id}/items": {
      "get": {
        "summary": "Get Folder Items",
        "description": "Every Movie + Series inside ONE library folder (folder-scoped poster wall).",
        "operationId": "get_folder_items_api_library_folders__folder_id__items_get",
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Folder Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderItemsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/library/series/{series_id}/episodes": {
      "get": {
        "summary": "Get Series Episodes",
        "description": "Every episode of one series, with per-episode playback facts (for TV).",
        "operationId": "get_series_episodes_api_library_series__series_id__episodes_get",
        "parameters": [
          {
            "name": "series_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Series Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/library/{item_id}/state": {
      "post": {
        "summary": "Set Item State",
        "description": "Mark an item watched/unwatched (roadmap item 2). Additive endpoint.",
        "operationId": "set_item_state_api_library__item_id__state_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemStateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/jellyfin/playback-info": {
      "get": {
        "summary": "Jellyfin Playback Info",
        "description": "List an item's audio + text-subtitle tracks for the player track pickers.",
        "operationId": "jellyfin_playback_info_api_jellyfin_playback_info_get",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jellyfin/subtitle": {
      "get": {
        "summary": "Jellyfin Subtitle",
        "description": "Proxy a Jellyfin **text** subtitle stream as WebVTT for the browser.\n\n``index`` is the subtitle stream's ``MediaStream.Index`` from playback-info.\nThe ``ms``/media source falls back to the item id (single-source items), so\n``/api/jellyfin/subtitle?id=<item>&ms=<source>&index=<n>`` just works.",
        "operationId": "jellyfin_subtitle_api_jellyfin_subtitle_get",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Id"
            }
          },
          {
            "name": "ms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "MediaSourceId (defaults to the item id)",
              "default": "",
              "title": "Ms"
            },
            "description": "MediaSourceId (defaults to the item id)"
          },
          {
            "name": "index",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Index"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FolderItemsResponse": {
        "properties": {
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "folder_id": {
            "type": "string",
            "title": "Folder Id"
          },
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "folder_id"
        ],
        "title": "FolderItemsResponse",
        "description": "GET /api/library/folders/{folder_id}/items \u2014 one folder's poster wall."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemStateRequest": {
        "properties": {
          "watched": {
            "type": "boolean",
            "title": "Watched",
            "default": true
          }
        },
        "type": "object",
        "title": "ItemStateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}